Move /view/<int:torrent_id> route into 'torrents' blueprint
and update templates.
