+++ nyaa/routes.py
@@ -1,11 +1,8 @@
-import json
 import os.path
 from urllib.parse import quote
 
 import flask
 from werkzeug.datastructures import CombinedMultiDict
-
-from sqlalchemy.orm import joinedload
 
 from nyaa import api_handler, app, backend, db, forms, models, template_utils, torrents, views
 from nyaa.utils import cached_function
@@ -58,72 +55,11 @@ def upload():
     if flask.request.method == 'POST' and upload_form.validate():
         torrent = backend.handle_torrent_upload(upload_form, flask.g.user)
 
-        return flask.redirect('/view/' + str(torrent.id))
+        return flask.redirect(flask.url_for('torrents.view', torrent_id=torrent.id))
     else:
         # If we get here with a POST, it means the form data was invalid: return a non-okay status
         status_code = 400 if flask.request.method == 'POST' else 200
         return flask.render_template('upload.html', upload_form=upload_form), status_code
-
-
-@app.route('/view/<int:torrent_id>', methods=['GET', 'POST'])
-def view_torrent(torrent_id):
-    if flask.request.method == 'POST':
-        torrent = models.Torrent.by_id(torrent_id)
-    else:
-        torrent = models.Torrent.query \
-                                .options(joinedload('filelist'),
-                                         joinedload('comments')) \
-                                .filter_by(id=torrent_id) \
-                                .first()
-    if not torrent:
-        flask.abort(404)
-
-    # Only allow admins see deleted torrents
-    if torrent.deleted and not (flask.g.user and flask.g.user.is_moderator):
-        flask.abort(404)
-
-    comment_form = None
-    if flask.g.user:
-        comment_form = forms.CommentForm()
-
-    if flask.request.method == 'POST':
-        if not flask.g.user:
-            flask.abort(403)
-
-        if comment_form.validate():
-            comment_text = (comment_form.comment.data or '').strip()
-
-            comment = models.Comment(
-                torrent_id=torrent_id,
-                user_id=flask.g.user.id,
-                text=comment_text)
-
-            db.session.add(comment)
-            db.session.flush()
-
-            torrent_count = torrent.update_comment_count()
-            db.session.commit()
-
-            flask.flash('Comment successfully posted.', 'success')
-
-            return flask.redirect(flask.url_for('view_torrent',
-                                                torrent_id=torrent_id,
-                                                _anchor='com-' + str(torrent_count)))
-
-    # Only allow owners and admins to edit torrents
-    can_edit = flask.g.user and (flask.g.user is torrent.user or flask.g.user.is_moderator)
-
-    files = None
-    if torrent.filelist:
-        files = json.loads(torrent.filelist.filelist_blob.decode('utf-8'))
-
-    report_form = forms.ReportForm()
-    return flask.render_template('view.html', torrent=torrent,
-                                 files=files,
-                                 comment_form=comment_form,
-                                 comments=torrent.comments,
-                                 can_edit=can_edit,
-                                 report_form=report_form)
 
 
 @app.route('/view/<int:torrent_id>/comment/<int:comment_id>/delete', methods=['POST'])
@@ -145,7 +81,7 @@ def delete_comment(torrent_id, comment_id):
     db.session.flush()
     torrent.update_comment_count()
 
-    url = flask.url_for('view_torrent', torrent_id=torrent.id)
+    url = flask.url_for('torrents.view', torrent_id=torrent.id)
     if flask.g.user.is_moderator:
         log = "Comment deleted on torrent [#{}]({})".format(torrent.id, url)
         adminlog = models.AdminLog(log=log, admin_id=flask.g.user.id)
@@ -196,7 +132,7 @@ def edit_torrent(torrent_id):
         if editor.is_moderator:
             torrent.deleted = form.is_deleted.data
 
-        url = flask.url_for('view_torrent', torrent_id=torrent.id)
+        url = flask.url_for('torrents.view', torrent_id=torrent.id)
         if deleted_changed and editor.is_moderator:
             log = "Torrent [#{0}]({1}) marked as {2}".format(
                 torrent.id, url, "deleted" if torrent.deleted else "undeleted")
@@ -276,7 +212,7 @@ def submit_report(torrent_id):
         db.session.commit()
         flask.flash('Successfully reported torrent!', 'success')
 
-    return flask.redirect(flask.url_for('view_torrent', torrent_id=torrent_id))
+    return flask.redirect(flask.url_for('torrents.view', torrent_id=torrent_id))
 
 
 def _get_cached_torrent_file(torrent):
@@ -304,6 +240,7 @@ def register_blueprints(flask_app):
     flask_app.register_blueprint(views.admin_bp)
     flask_app.register_blueprint(views.main_bp)
     flask_app.register_blueprint(views.site_bp)
+    flask_app.register_blueprint(views.torrents_bp)
     flask_app.register_blueprint(views.users_bp)
