+++ nyaa/views/__init__.py
@@ -3,6 +3,7 @@ from nyaa.views import (
     admin,
     main,
     site,
+    torrents,
     users,
 )
 
@@ -10,4 +11,5 @@ account_bp = account.bp
 admin_bp = admin.bp
 main_bp = main.bp
 site_bp = site.bp
+torrents_bp = torrents.bp
 users_bp = users.bp
